# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AnalyticsOverview:
    properties:
      this_month:
        type: integer
      this_week:
        type: integer
      today:
        type: integer
      total_clicks:
        type: integer
    type: object
...
  models.BrowserAnalytics:
    properties:
      browser:
        type: string
      count:
        type: integer
    type: object
  models.DailyClick:
    properties:
      clicks:
        type: integer
      date:
        type: string
    type: object
  models.DeviceAnalytics:
    properties:
      count:
        type: integer
      device:
        type: string
    type: object
  models.HTTPDetailsError:
    properties:
      details:
        example: connection refused
        type: string
      error:
        example: failed to fetch analytics
        type: string
    type: object
  models.HTTPError:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
  models.HTTPResponseErr:
    properties:
      err:
        example: not found
        type: string
    type: object
...
  models.PaginatedURLResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.URL'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
  models.PaginatedUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.User'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
  models.PaginationMeta:
    properties:
      limit:
        example: 50
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
  models.RecentVisit:
    properties:
      browser:
        type: string
      created_at:
        type: string
      device:
        type: string
      ip_address:
        type: string
      referer:
        type: string
    type: object
  models.URL:
    properties:
      clicks:
        type: integer
      created_at:
        type: string
      expiry:
        type: string
      id:
        description: go for uuid later
        type: integer
      long_url:
        type: string
      short_url:
        type: string
      user_id:
        type: integer
    required:
    - long_url
    - short_url
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: High performance URL shortener API with cache-aside caching and analytics.
  title: URL Shortener
  version: "0.67"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all users in the system (Admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedUserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPResponseErr'
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - admin
  /admin/users/{userid}/urls:
    get:
      consumes:
      - application/json
      description: Get a paginated list of URLs belonging to a specific user (Admin
        only)
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPResponseErr'
      security:
      - Bearer: []
      summary: Get URLs of a specific user
      tags:
      - admin
  /analytics/{id}/browser:
    get:
      description: Get visitor distribution stats grouped by browser type for a URL
        ID
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BrowserAnalytics'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPDetailsError'
      security:
      - Bearer: []
      summary: Get browser distribution analytics
      tags:
      - analytics
  /analytics/{id}/daily:
    get:
      description: Get a list of dates and click counts for a URL ID
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DailyClick'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPDetailsError'
      security:
      - Bearer: []
      summary: Get daily click analytics
      tags:
      - analytics
  /analytics/{id}/device:
    get:
      description: Get visitor distribution stats grouped by device type for a URL
        ID
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeviceAnalytics'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPDetailsError'
      security:
      - Bearer: []
      summary: Get device distribution analytics
      tags:
      - analytics
  /analytics/{id}/overview:
    get:
      description: Get click stats summary (today, this week, this month, total clicks)
        for a URL ID
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnalyticsOverview'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPDetailsError'
      security:
      - Bearer: []
      summary: Get URL analytics overview
      tags:
      - analytics
  /analytics/{id}/recent:
    get:
      description: Get up to 20 recent visitor events (IP, User Agent, Referer, Browser,
        Device, CreatedAt) for a URL ID
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RecentVisit'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPDetailsError'
      security:
      - Bearer: []
      summary: Get recent visits logs
      tags:
      - analytics
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"